requirements.txt
# Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# Machine Learning
scikit-learn>=1.3.0
joblib>=1.3.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0

# Image Processing
opencv-python>=4.8.0
albumentations>=1.3.0
Pillow>=10.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Progress bars
tqdm>=4.65.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.25.0
